requirements.txt
aiofiles==23.1.0
amqp==5.1.1
anyio==3.6.2
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
base2048==0.1.3
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.0.1
ci-info==0.3.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
configobj==5.0.8
configparser==5.3.0
cryptography==39.0.2
defusedxml==0.7.1
deprecation==2.1.0
diff-match-patch==20200713
Django==4.0.8
django-cors-headers==3.13.0
django-eav2==1.2.3
django-filter==22.1
django-import-export==3.0.1
django-jsonfield==1.4.1
django-jsonfield-backport==1.0.5
django-otp==1.1.3
django-reset-migrations==0.4.0
django-rest-framework==0.1.0
django-rql==4.2.3
django-session-timeout==0.1.0
django-simple-history==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docx-mailmerge==0.5.0
docx-mailmerge2==0.6.2
docx2pdf==0.1.8
et-xmlfile==1.1.0
etelemetry==0.3.0
filelock==3.9.1
fitz==0.0.1.dev2
fpdf==1.7.2
frontend==0.0.3
future==0.18.3
h11==0.14.0
httplib2==0.21.0
idna==3.4
isodate==0.6.1
itsdangerous==2.1.2
kombu==5.2.4
lark-parser==0.11.0
lib-rql==1.1.7
looseversion==1.1.2
lxml==4.9.2
MarkupPy==1.14
msgpack==1.0.5
networkx==3.0
nibabel==5.0.1
nipype==1.8.5
numpy==1.24.2
odfpy==1.4.1
openpyxl==3.0.10
packaging==23.0
pandas==1.5.3
pathlib==1.0.1
pdfminer.six==20221105
PDFNetPython3==9.5.0.post1
pdfrw==0.4
pikepdf==7.1.1
Pillow==9.4.0
prompt-toolkit==3.0.38
prov==2.0.0
psycopg2==2.9.5
py==1.4.22
pycparser==2.21
PyDocX==0.9.10
pydot==1.4.2
PyJWT==2.6.0
pyparsing==3.0.9
PyPDF2==3.0.1
pytest==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-docx==0.8.11
python-http-client==3.3.7
pytz==2022.5
pyxnat==1.5
PyYAML==6.0
rdflib==6.2.0
redis==4.5.1
reportlab==3.6.12
requests==2.28.2
scipy==1.10.1
sendgrid==6.9.7
simplejson==3.18.4
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
starlette==0.26.1
traits==6.3.2
typing_extensions==4.5.0
uno==0.3.3
urllib3==1.26.14
uvicorn==0.21.0
vine==5.0.0
wcwidth==0.2.6
xlrd==2.0.1
xlwt==1.3.0
